Cleanup serial communication
There's a lot of communication being sent over the Serial
bus that isn't related to the operation of the trailer. Disable
blocks of debug code that were printing human readable messages
to the Serial console. They can be turned back on if needed by
uncommenting the _SERIAL_DEBUG definition at the top of the file.